+++ src/Maestro/API/V1/General.hs
@@ -13,11 +13,11 @@ data GeneralAPI route =
         :> Get '[JSON] TimestampedSystemStart
     , eraHistory
         :: route
-        :- "era-history"
+        :- "era-summaries"
         :> Get '[JSON] TimestampedEraSummaries
     , protocolParams
         :: route
-        :- "protocol-params"
+        :- "protocol-parameters"
         :> Get '[JSON] TimestampedProtocolParameters
     , chainTip
         :: route