+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Revision history for `maestro-sdk`
 
+## [1.7.0](https://github.com/maestro-org/haskell-sdk/compare/v1.5.0..v1.6.0) -- 2024-08-17
+
+* Updated to newer `era-summaries` and `protocol-parameters` endpoint from earlier `era-history` and `protocol-params` respectively.
+* `AssetInfo` type which corresponds to `/assets/:asset` is updated with more response fields.
+
 ## [1.6.0](https://github.com/maestro-org/haskell-sdk/compare/v1.5.0..v1.6.0) -- 2024-04-09
 
 Added: